+++ gillespy2/remote/client/compute_server.py
@@ -17,7 +17,7 @@ ComputeServer(Server)
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from stochss_compute.client.server import Server
+from gillespy2.remote.client.server import Server
 
 class ComputeServer(Server):
     '''
