+++ gillespy2/remote/core/errors.py
@@ -1,5 +1,5 @@
 '''
-stochss_compute.core.errors
+gillespy2.remote.core.errors
 '''
 # StochSS-Compute is a tool for running and caching GillesPy2 simulations remotely.
 # Copyright (C) 2019-2023 GillesPy2 and StochSS developers.
