+++ gillespy2/remote/server/__init__.py
@@ -1,5 +1,5 @@
 '''
-stochss_compute.server
+gillespy2.remote.server
 '''
 # StochSS-Compute is a tool for running and caching GillesPy2 simulations remotely.
 # Copyright (C) 2019-2023 GillesPy2 and StochSS developers.
