+++ gillespy2/remote/server/run.py
@@ -1,5 +1,5 @@
 '''
-stochss_compute.server.run
+gillespy2.remote.server.run
 '''
 # StochSS-Compute is a tool for running and caching GillesPy2 simulations remotely.
 # Copyright (C) 2019-2023 GillesPy2 and StochSS developers.
@@ -17,29 +17,37 @@ stochss_compute.server.run
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import random
 from datetime import datetime
-from secrets import token_hex
 
 from tornado.web import RequestHandler
 from tornado.ioloop import IOLoop
-from distributed import Client, Future
-from gillespy2.core import Results
-from stochss_compute.core.messages.status import SimStatus
-from stochss_compute.core.messages.simulation_run import SimulationRunRequest, SimulationRunResponse
-from stochss_compute.server.cache import Cache
+from distributed import Client
+from gillespy2.remote.core.messages.status import SimStatus
+from gillespy2.remote.core.exceptions import PRNGCollision
+from gillespy2.remote.core.messages.simulation_run import SimulationRunRequest, SimulationRunResponse
+from gillespy2.remote.server.cache import Cache
 
+from gillespy2.remote.core.log_config import init_logging
+log = init_logging(__name__)
 
-class RunHandler(RequestHandler):
+class SimulationRunHandler(RequestHandler):
     '''
     Endpoint for running Gillespy2 simulations.
     '''
 
-    scheduler_address = None
-    cache_dir = None
+    def __init__(self, application, request, **kwargs):
+        self.scheduler_address = None
+        self.cache_dir = None
+        self.key = None
+        super().__init__(application, request, **kwargs)
+
+    def data_received(self, chunk: bytes):
+        raise NotImplementedError()
+    
     def initialize(self, scheduler_address, cache_dir):
         '''
         Sets the address to the Dask scheduler and the cache directory.
+        Creates a new directory for one-off results files identifiable by token.
 
         :param scheduler_address: Scheduler address.
         :type scheduler_address: str
@@ -47,70 +55,75 @@ class RunHandler(RequestHandler):
         :param cache_dir: Path to the cache.
         :type cache_dir: str
         '''
-
         self.scheduler_address = scheduler_address
-        self.cache_dir = cache_dir
+        while cache_dir.endswith('/'):
+            cache_dir = cache_dir[:-1]
+        self.cache_dir = cache_dir + '/run/'
 
     async def post(self):
         '''
-        Process simulation run request.
+        Process simulation run  POST request.
         '''
         sim_request = SimulationRunRequest.parse(self.request.body)
-        sim_hash = sim_request.hash()
-        log_string = f'{datetime.now()} | <{self.request.remote_ip}> | Simulation Run Request | <{sim_hash}> | '
-        cache = Cache(self.cache_dir, sim_hash)
-        if not cache.exists():
-            cache.create()
-        empty = cache.is_empty()
-        if not empty:
-            # Check the number of trajectories in the request, default 1
-            n_traj = sim_request.kwargs.get('number_of_trajectories', 1)
-            # Compare that to the number of cached trajectories
-            trajectories_needed =  cache.n_traj_needed(n_traj)
-            if trajectories_needed > 0:
-                sim_request.kwargs['number_of_trajectories'] = trajectories_needed
-                print(log_string +
-                    f'Partial cache. Running {trajectories_needed} new trajectories.')
-                client = Client(self.scheduler_address)
-                future = self._submit(sim_request, sim_hash, client)
-                self._return_running(sim_hash, future.key)
-                IOLoop.current().run_in_executor(None, self._cache, sim_hash, future, client)
-            else:
-                print(log_string + 'Returning cached results.')
-                results = cache.get()
-                ret_traj = random.sample(results, n_traj)
-                new_results = Results(ret_traj)
-                new_results_json = new_results.to_json()
-                sim_response = SimulationRunResponse(SimStatus.READY, results_id = sim_hash, results = new_results_json)
-                self.write(sim_response.encode())
-                self.finish()
-        if empty:
-            print(log_string + 'Results not cached. Running simulation.')
-            client = Client(self.scheduler_address)
-            future = self._submit(sim_request, sim_hash, client)
-            self._return_running(sim_hash, future.key)
-            IOLoop.current().run_in_executor(None, self._cache, sim_hash, future, client)
+        self.key = sim_request.key
+        cache = Cache(self.cache_dir, self.key)
+        if cache.exists():
+            self.set_status(404, reason='Try again with a different key, because that one is taken.')
+            self.finish()
+            raise PRNGCollision('Try again with a different key, because that one is taken.')
+        cache.create()
+        client = Client(self.scheduler_address)
+        future = self._submit(sim_request, client)
+        log_string = f'{datetime.now()} | <{self.request.remote_ip}> | Simulation Run  Request | <{self.key}> | '
+        print(log_string + 'Running simulation.')
+        self._return_running()
+        IOLoop.current().run_in_executor(None, self._cache, future, client)
 
-    def _cache(self, sim_hash, future: Future, client: Client):
+    def _cache(self, future, client):
+        '''
+        Await results, close client, save to disk.
+
+        :param future: Handle to the running simulation, to be awaited upon.
+        :type future: distributed.Future
+
+        :param client: Client to the Dask scheduler. Closing here for good measure, not sure if strictly necessary.
+        :type client: distributed.Client
+        '''
         results = future.result()
         client.close()
-        cache = Cache(self.cache_dir, sim_hash)
+        cache = Cache(self.cache_dir, self.key)
         cache.save(results)
 
-    def _submit(self, sim_request, sim_hash, client: Client):
+    def _submit(self, sim_request, client):
+        '''
+        Submit request to dask scheduler.
+        Uses pydoc.locate to convert str to solver class name object.
+
+        :param sim_request: The user's request for a  simulation.
+        :type sim_request: SimulationRunRequest
+
+        :param client: Client to the Dask scheduler.
+        :type client: distributed.Client
+
+        :returns: Handle to the running simulation and the results on the worker.
+        :rtype: distributed.Future
+        '''
         model = sim_request.model
         kwargs = sim_request.kwargs
-        n_traj = kwargs.get('number_of_trajectories', 1)
+        key = sim_request.key
         if "solver" in kwargs:
+            # pylint:disable=import-outside-toplevel
             from pydoc import locate
+            # pylint:enable=import-outside-toplevel
             kwargs["solver"] = locate(kwargs["solver"])
 
-        # keep client open for now! close?
-        key = f'{sim_hash}:{n_traj}:{token_hex(8)}'
         future = client.submit(model.run, **kwargs, key=key)
         return future
 
-    def _return_running(self, results_id, task_id):
-        sim_response = SimulationRunResponse(SimStatus.RUNNING, results_id=results_id, task_id=task_id)
+    def _return_running(self):
+        '''
+        Let the user know we submitted the simulation to the scheduler.
+        '''
+        sim_response = SimulationRunResponse(SimStatus.RUNNING)
         self.write(sim_response.encode())
         self.finish()
